fix(VST-231): preflight rejects schema-invalid source agent-frontmatter
`MappingConfig` marks both `agent_frontmatter` fields `#[serde(skip)]`, so
deserializing a source `vstack.toml` as a `MappingConfig` never looked at
`[agent-frontmatter]`; `parse_agent_frontmatter_tables`, which does, dropped
every entry it could not interpret. A source whose override was a non-table or
carried a wrongly typed field passed the preflight, refreshed the agent without
the override, and recorded the changed source hash as satisfied.

The parse is now a walk that reports each dropped entry, with
`validate_agent_frontmatter_tables` returning the first reason and
`parse_agent_frontmatter_tables` ignoring them — validation cannot drift from
what the parse accepts. The refresh preflight fails closed on that reason.
Unknown fields inside a well-formed override stay accepted; every real consumer
`vstack.toml` validates unchanged.

Splits `refresh_sources/tests.rs` by subject under test: source resolution
stays, the remote-cache clone/update/validate suite moves to
`refresh_sources/tests/remote_cache.rs`, sharing the parent's helpers. All 38
tests still run.
